significatantly simplified keywords and grammar
keywords are now only prefix. ## for infix functions, including builtins / std. @@ for macros like @@open but dynamic import would be std and be #import. Not trying to overly simplify the grammar but ## i a bit ugly/verbose which is the point because Si prefers Lisp-style S-expressions over infix function calls